Module northstar_runtime::npk::npk

NPK file format

Structs

Error

Npk loading Error

Hashes

NPK Hashes

Meta

NPK archive comment

Npk

Northstar package

SquashfsOptions

Squashfs Options

Enums

CompressionAlgorithm

Squashfs compression algorithm

Constants

FS_IMG_NAME

File system file name

MANIFEST_NAME

Manifest file name

MKSQUASHFS

Default path to mksquashfs

NPK_EXT

NPK extension

SIGNATURE_NAME

Signature file name

UNSQUASHFS

Default path to unsquashfs

Functions

generate_key

Generate a keypair suitable for signing and verifying NPKs

pack

Create an NPK for the northstar runtime. northstar-sextant collects the artifacts in a given container directory, creates and signs the necessary metadata and packs the results into a zipped NPK file.

pack_with

Create an NPK with special squashfs options northstar-sextant collects the artifacts in a given container directory, creates and signs the necessary metadata and packs the results into a zipped NPK file.

unpack

Extract the npk content to out

unpack_with

Extract the npk content to out with a give unsquashfs binary